Improved independence of aperture base classes
