+++ pygerber/meta/arc_util_mixin.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import annotations
 
-from math import cos, degrees, radians, sin, tau
+from math import cos, degrees, radians, sin
 
 from pygerber.mathclasses import Vector2D, angle_from_zero
 from pygerber.meta.spec import ArcSpec
@@ -29,12 +29,15 @@ class ArcUtilMixin:
         begin_angle, end_angle = self.get_begin_end_angles(spec)
         radius = spec.get_radius()
         x, y = self.get_arc_co_functions(radius)
-        delta = self._get_delta_angle(begin_angle, end_angle, radius)
+        delta = self.get_arc_traverse_step_angle(begin_angle, end_angle, radius)
         if self.isCCW:
             return self._get_arc_points_ccw(end_angle, begin_angle, x, spec, y, delta)
         else:
             return self._get_arc_points_cw(end_angle, begin_angle, x, spec, y, delta)
 
+    def get_arc_traverse_step_angle(self, begin_angle, end_angle, radius):
+        raise NotImplementedError("get_arc_traverse_step_angle() have to be implemented in subclass.")
+        
     def _get_arc_points_ccw(self, end_angle, begin_angle, x, spec, y, delta):
         end_relative_angle = end_angle - begin_angle
         angle_offset = begin_angle
@@ -56,25 +59,6 @@ class ArcUtilMixin:
                 y(current_angle + angle_offset) + spec.center.y,
             )
             current_angle -= delta
-
-    def _get_delta_angle(self, begin_angle, end_angle, radius):
-        relative_angle = self.get_relative_angle(begin_angle, end_angle)
-        arc_ratio = self.get_arc_ratio(relative_angle)
-        arc_length = self.get_arc_length(radius) * arc_ratio
-        number_of_points = int(arc_length * self.dpmm)
-        return relative_angle / number_of_points
-
-    @staticmethod
-    def get_arc_length(radius) -> float:
-        return tau * radius
-
-    @staticmethod
-    def get_arc_ratio(relative_angle):
-        return relative_angle / 360
-
-    @staticmethod
-    def get_relative_angle(begin_angle, end_angle):
-        return end_angle - begin_angle
 
     @staticmethod
     def get_arc_co_functions(radius):
