+++ pygerber/parser/pillow/apertures/polygon.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import annotations
+from pygerber.parser.pillow.apertures.arc_mixin import ArcUtilMixinPillow
 
 from PIL import Image, ImageDraw
 from pygerber.meta.aperture import PolygonAperture
@@ -7,7 +8,7 @@ from pygerber.meta.spec import ArcSpec, FlashSpec, LineSpec
 from pygerber.parser.pillow.apertures.util import PillowUtilMethdos
 
 
-class PillowPolygon(PolygonAperture, PillowUtilMethdos):
+class PillowPolygon(ArcUtilMixinPillow, PolygonAperture, PillowUtilMethdos):
     draw_canvas: ImageDraw.ImageDraw
 
     def flash(self, spec: FlashSpec) -> None:
